Add files via upload
I'm using VSCode and this is what I have so far. When I run the index.html with the live server with the VSCode extension, I just get a black square with a smaller green square within. I am trying to create a web-based game for a student innovation school project and need help. I am a beginner with coding and have very limited knowledge and skill. Any assistance would be greatly appreciated.